--- src/main/java/com/trevari/test/domain/book/dto/BookSearchDto.java
@@ -1,9 +1,0 @@
-package com.trevari.test.domain.book.dto;
-
-public record BookSearchDto(
-        Long isbn
-) {
-    public static BookSearchDto of(String id) {
-        return new BookSearchDto(Long.parseLong(id));
-    }
-}
