--- src/main/java/com/trevari/test/domain/book/dto/BooksSearchKeywordDto.java
@@ -1,12 +1,0 @@
-package com.trevari.test.domain.book.dto;
-
-import org.springframework.data.domain.Pageable;
-
-public record BooksSearchKeywordDto(
-        String keyword,
-        Pageable pageable
-) {
-    public static BooksSearchKeywordDto of(String keyword, Pageable pageable) {
-        return new BooksSearchKeywordDto(keyword, pageable);
-    }
-}
