+++ src/main/java/com/trevari/test/domain/book/infrastructure/repository/BookRepository.java
@@ -1,0 +1,10 @@
+package com.trevari.test.domain.book.infrastructure.repository;
+
+import com.trevari.test.domain.book.entity.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
