+++ src/main/java/com/trevari/test/domain/book/port/in/SearchBookUseCase.java
@@ -1,0 +1,4 @@
+package com.trevari.test.domain.book.port.in;
+
+public interface SearchBookUseCase {
+}
