+++ src/main/java/com/trevari/test/domain/book/port/in/dto/BookSearchDto.java
@@ -1,0 +1,9 @@
+package com.trevari.test.domain.book.port.in.dto;
+
+public record BookSearchDto(
+        Long isbn
+) {
+    public static BookSearchDto of(String id) {
+        return new BookSearchDto(Long.parseLong(id));
+    }
+}
