--- src/main/java/com/trevari/test/domain/book/repository/BookRepository.java
@@ -1,10 +1,0 @@
-package com.trevari.test.domain.book.repository;
-
-import com.trevari.test.domain.book.entity.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface BookRepository extends JpaRepository<Book, Long> {
-}
