+++ src/test/java/com/trevari/test/domain/book/seed/service/SeedServiceTest.java
@@ -1,6 +1,6 @@
 package com.trevari.test.domain.book.seed.service;
 
-import com.trevari.test.domain.book.repository.BookRepository;
+import com.trevari.test.domain.book.infrastructure.repository.BookRepository;
 import com.trevari.test.domain.book.seed.dto.SeedDetailResponse;
 import com.trevari.test.domain.book.seed.dto.SeedSearchResponse;
 import org.junit.jupiter.api.BeforeEach;
